requirements/production.txt
# Pro-tip: Try not to put anything here. Avoid dependencies in
#	production that aren't in development.
-r base.txt



# WSGI Handler
# ------------------------------------------------
gevent==1.2.2
gunicorn==19.7.1

# Static and Media Storage
# ------------------------------------------------
django-storages-redux==1.3.3

# Email backends for Mailgun, Postmark, SendGrid and more
# -------------------------------------------------------
django-cors-headers
raven
paraer
django-rest-framework-social-oauth2
